+++ omc3/sbs_propagation.py
@@ -65,7 +65,7 @@ from omc3.model.constants import (
 )
 from omc3.model.model_creators.manager import CreatorType, get_model_creator_class
 from omc3.segment_by_segment.constants import logfile
-from omc3.segment_by_segment.propagables import Propagable, ALL_PROPAGABLES 
+from omc3.segment_by_segment.propagables import AlphaPhase, BetaPhase, Phase, Propagable, F1001, F1010, ALL_PROPAGABLES 
 from omc3.segment_by_segment.segments import (
     SbsDefinitionError,
     Segment,
@@ -206,7 +206,14 @@ def create_segment(
         f"This has been input as {segment_in!s}."
     )
 
-    propagables = [propg(segment, measurement, accel.elements) for propg in ALL_PROPAGABLES]
+    measured_classes: list[type[Propagable]] = [propg for propg in ALL_PROPAGABLES if propg.in_measurement(measurement)]
+    if Phase not in measured_classes:
+        raise FileNotFoundError("No measurement for Phase found in the given measurement directory.")
+    
+    # if (F1001 in measured_classes or F1010 in measured_classes) and BetaPhase not in measured_classes:
+    #     raise FileNotFoundError("Beta Phase is required for F1001 and F1010 propagation.")
+    
+    propagables: list[Propagable] = [propg(segment, measurement, accel.elements) for propg in measured_classes]
 
     # Create the segment via madx
     creator_class = get_model_creator_class(accel, CreatorType.SEGMENT)