+++ omc3/segment_by_segment/propagables/dispersion.py
@@ -41,6 +41,16 @@ class Dispersion(Propagable):
         error = meas.dispersion[plane].loc[names, c.error_column]
         return dispersion, error
     
+    @classmethod
+    def in_measurement(cls, meas: OpticsMeasurement) -> bool:
+        """ Check if the dispersion is in the measurement data. """
+        try:
+            meas.dispersion_x
+            meas.dispersion_y
+        except FileNotFoundError:
+            return False
+        return True
+    
     def get_segment_observation_points(self, plane: str):
         """ Return the measurement points for the given plane, that are in the segment. """
         return common_indices(